feat: changes code from view to ocntainer
