+++ src/components/container/PlayerSearchContainer.js
@@ -1,12 +1,32 @@
 import React, { Component } from 'react'
 import PlayerSearchView from '../views/PlayerSearchView'
+import {connect} from 'react-redux'
+import { fetchPlayerInfoThunk } from '../../thunks'
 
-export default class PlayerSearchContainer extends Component {
+
+class PlayerSearchContainer extends Component {
+  componentDidMount() {
+    this.props.fetchPlayerInfo()
+    console.log('componentDidMount')
+  }
+
   render() {
     return (
       <div>
-        <PlayerSearchView />
+        <PlayerSearchView allPlayers={this.props.allPlayers}/>
       </div>
     )
   }
 }
+
+const mapStateToProps = state => {
+  console.log({state})
+}
+
+const mapDispatchToProps = dispatch => {
+  return {
+    fetchPlayerInfo: () => dispatch(fetchPlayerInfoThunk())
+  }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(PlayerSearchContainer)
