+++ src/components/views/PlayerSearchView.js
@@ -1,14 +1,7 @@
 import React, { Component } from 'react'
 import SearchFieldView from './SearchFieldView'
-import {connect} from 'react-redux'
-import { fetchPlayerInfoThunk } from '../../thunks'
 
 class PlayerSearchView extends Component {
-  componentDidMount() {
-    this.props.fetchPlayerInfo()
-    console.log('componentDidMount')
-  }
-
   render() {
     return (
       <div>
@@ -18,14 +11,4 @@ class PlayerSearchView extends Component {
   }
 }
 
-const mapStateToProps = state => {
-  console.log({state})
-}
-
-const mapDispatchToProps = dispatch => {
-  return {
-    fetchPlayerInfo: () => dispatch(fetchPlayerInfoThunk())
-  }
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(PlayerSearchView)
+export default PlayerSearchView
