+++ src/thunks/index.js
@@ -1,2 +1,2 @@
 // Barrel file for thunks, which will be mapped to and invoked within our smart containers;
-export * from '../store/utilities/PlayerSearch';
+export * from '../store/utilities/playerSearch';
